# PocCompassFunc/openapi.json
{
 "components": {
  "schemas": {
...
   "AssumptionSeed": {
    "properties": {
     "priority": {
      "$ref": "#/components/schemas/Priority",
      "default": "medium"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "AssumptionSeed",
    "type": "object"
   },
...
   "CardLabels": {
    "properties": {
     "deadline_risk": {
      "title": "Deadline Risk",
      "type": "string"
     },
     "deadline_risk_hint": {
      "title": "Deadline Risk Hint",
      "type": "string"
     },
     "drift": {
      "title": "Drift",
      "type": "string"
     },
     "drift_hint": {
      "title": "Drift Hint",
      "type": "string"
     },
     "under_evidenced_title": {
      "title": "Under Evidenced Title",
      "type": "string"
     },
     "unnecessary": {
      "title": "Unnecessary",
      "type": "string"
     },
     "unnecessary_hint": {
      "title": "Unnecessary Hint",
      "type": "string"
     },
     "untested": {
      "title": "Untested",
      "type": "string"
     },
     "untested_hint": {
      "title": "Untested Hint",
      "type": "string"
     },
     "weak_tasks_title": {
      "title": "Weak Tasks Title",
      "type": "string"
     }
    },
    "required": [
     "drift",
     "drift_hint",
     "unnecessary",
     "unnecessary_hint",
     "deadline_risk",
     "deadline_risk_hint",
     "untested",
     "untested_hint",
     "weak_tasks_title",
     "under_evidenced_title"
    ],
    "title": "CardLabels",
    "type": "object"
   },
   "Cards": {
    "properties": {
     "deadline_risk": {
      "title": "Deadline Risk",
      "type": "integer"
     },
     "drift": {
      "title": "Drift",
      "type": "integer"
     },
     "unnecessary": {
      "title": "Unnecessary",
      "type": "integer"
     },
     "untested": {
      "title": "Untested",
      "type": "integer"
     }
    },
    "required": [
     "drift",
     "unnecessary",
     "deadline_risk",
     "untested"
    ],
    "title": "Cards",
    "type": "object"
   },
...
   "CriterionSeed": {
    "properties": {
     "target": {
      "default": "",
      "maxLength": 200,
      "title": "Target",
      "type": "string"
     },
     "text": {
      "maxLength": 1000,
      "minLength": 1,
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "CriterionSeed",
    "type": "object"
   },
...
   "Dashboard": {
    "properties": {
     "alerts": {
      "items": {
       "$ref": "#/components/schemas/Alert"
      },
      "title": "Alerts",
      "type": "array"
     },
     "cards": {
      "$ref": "#/components/schemas/Cards"
     },
     "deadline_risks": {
      "items": {
       "$ref": "#/components/schemas/DeadlineRiskItem"
      },
      "title": "Deadline Risks",
      "type": "array"
     },
     "design_changed": {
      "title": "Design Changed",
      "type": "boolean"
     },
     "health": {
      "$ref": "#/components/schemas/HealthView"
     },
     "last_evaluated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Evaluated At"
     },
     "schedule": {
      "$ref": "#/components/schemas/ScheduleView"
     },
     "stale_task_ids": {
      "items": {
       "type": "string"
      },
      "title": "Stale Task Ids",
      "type": "array"
     },
     "trend": {
      "items": {
       "$ref": "#/components/schemas/TrendPoint"
      },
      "title": "Trend",
      "type": "array"
     },
     "under_evidenced": {
      "items": {
       "$ref": "#/components/schemas/UnderEvidenced"
      },
      "title": "Under Evidenced",
      "type": "array"
     },
     "unevaluated_task_ids": {
      "items": {
       "type": "string"
      },
      "title": "Unevaluated Task Ids",
      "type": "array"
     },
     "weak_tasks": {
      "items": {
       "$ref": "#/components/schemas/WeakTask"
      },
      "title": "Weak Tasks",
      "type": "array"
     }
    },
    "required": [
     "health",
     "cards",
     "schedule",
     "weak_tasks",
     "under_evidenced",
     "deadline_risks",
     "alerts",
     "trend",
     "stale_task_ids",
     "unevaluated_task_ids",
     "design_changed",
     "last_evaluated_at"
    ],
    "title": "Dashboard",
    "type": "object"
   },
...
   "Evaluation": {
    "properties": {
     "ai_calls": {
      "default": 0,
      "title": "Ai Calls",
      "type": "integer"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "design_hash": {
      "title": "Design Hash",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     },
     "task_results": {
      "items": {
       "$ref": "#/components/schemas/TaskResult"
      },
      "title": "Task Results",
      "type": "array"
     },
     "trigger": {
      "$ref": "#/components/schemas/EvaluationTrigger"
     },
     "unevaluated_task_ids": {
      "items": {
       "type": "string"
      },
      "title": "Unevaluated Task Ids",
      "type": "array"
     }
    },
    "required": [
     "id",
     "projectId",
     "created_at",
     "trigger",
     "model",
     "design_hash",
     "task_results",
     "ai_calls",
     "unevaluated_task_ids"
    ],
    "title": "Evaluation",
    "type": "object"
   },
   "EvaluationTrigger": {
    "enum": [
     "manual",
     "scheduled"
    ],
    "title": "EvaluationTrigger",
    "type": "string"
   },
...
   "ExtractRequest": {
    "properties": {
     "text": {
      "minLength": 1,
      "title": "Text",
      "type": "string"
     }
    },
    "required": [
     "text"
    ],
    "title": "ExtractRequest",
    "type": "object"
   },
   "ExtractResponse": {
    "properties": {
     "tasks": {
      "items": {
       "$ref": "#/components/schemas/ExtractedTask"
      },
      "title": "Tasks",
      "type": "array"
     }
    },
    "required": [
     "tasks"
    ],
    "title": "ExtractResponse",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Mode": {
    "enum": [
     "poc",
     "planning",
     "improvement"
    ],
    "title": "Mode",
    "type": "string"
   },
   "ModeDefinition": {
    "properties": {
     "card_labels": {
      "$ref": "#/components/schemas/CardLabels"
     },
     "description": {
      "maxLength": 300,
      "title": "Description",
      "type": "string"
     },
     "id": {
      "$ref": "#/components/schemas/Mode"
     },
     "labels": {
      "$ref": "#/components/schemas/ModeLabels"
     },
     "name": {
      "maxLength": 30,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "placeholders": {
      "$ref": "#/components/schemas/Placeholders"
     },
     "prompt_guidance": {
      "maxLength": 4000,
      "title": "Prompt Guidance",
      "type": "string"
     },
     "updated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated At"
     },
     "updated_by": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Updated By"
     },
     "weights": {
      "$ref": "#/components/schemas/Weights"
     }
    },
    "required": [
     "name",
     "description",
     "labels",
     "card_labels",
     "placeholders",
     "weights",
     "prompt_guidance",
     "id",
     "updated_at",
     "updated_by"
    ],
    "title": "ModeDefinition",
    "type": "object"
   },
   "ModeDefinitionBody": {
    "properties": {
     "card_labels": {
      "$ref": "#/components/schemas/CardLabels"
     },
     "description": {
      "maxLength": 300,
      "title": "Description",
      "type": "string"
     },
     "labels": {
      "$ref": "#/components/schemas/ModeLabels"
     },
     "name": {
      "maxLength": 30,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "placeholders": {
      "$ref": "#/components/schemas/Placeholders"
     },
     "prompt_guidance": {
      "maxLength": 4000,
      "title": "Prompt Guidance",
      "type": "string"
     },
     "weights": {
      "$ref": "#/components/schemas/Weights"
     }
    },
    "required": [
     "name",
     "description",
     "labels",
     "card_labels",
     "placeholders",
     "weights",
     "prompt_guidance"
    ],
    "title": "ModeDefinitionBody",
    "type": "object"
   },
   "ModeLabels": {
    "properties": {
     "assumption": {
      "title": "Assumption",
      "type": "string"
     },
     "criterion": {
      "title": "Criterion",
      "type": "string"
     },
     "deadline": {
      "title": "Deadline",
      "type": "string"
     },
     "decision_continue": {
      "title": "Decision Continue",
      "type": "string"
     },
     "decision_pivot": {
      "title": "Decision Pivot",
      "type": "string"
     },
     "decision_stop": {
      "title": "Decision Stop",
      "type": "string"
     },
     "evidence": {
      "title": "Evidence",
      "type": "string"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "task": {
      "title": "Task",
      "type": "string"
     }
    },
    "required": [
     "goal",
     "assumption",
     "criterion",
     "deadline",
     "task",
     "evidence",
     "decision_continue",
     "decision_pivot",
     "decision_stop"
    ],
    "title": "ModeLabels",
    "type": "object"
   },
...
   "Placeholders": {
    "properties": {
     "assumption": {
      "title": "Assumption",
      "type": "string"
     },
     "criterion": {
      "title": "Criterion",
      "type": "string"
     },
     "evidence": {
      "title": "Evidence",
      "type": "string"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "task": {
      "title": "Task",
      "type": "string"
     },
     "title": {
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "title",
     "goal",
     "assumption",
     "criterion",
     "task",
     "evidence"
    ],
    "title": "Placeholders",
    "type": "object"
   },
...
   "Project": {
    "properties": {
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "deadline": {
      "format": "date",
      "title": "Deadline",
      "type": "string"
     },
     "deleted": {
      "default": false,
      "title": "Deleted",
      "type": "boolean"
     },
     "goal": {
      "title": "Goal",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "members": {
      "items": {
       "type": "string"
      },
      "title": "Members",
      "type": "array"
     },
     "mode": {
      "$ref": "#/components/schemas/Mode"
     },
     "owner_email": {
      "title": "Owner Email",
      "type": "string"
     },
     "start_date": {
      "format": "date",
      "title": "Start Date",
      "type": "string"
     },
     "status": {
      "$ref": "#/components/schemas/ProjectStatus",
      "default": "active"
     },
     "summary": {
      "$ref": "#/components/schemas/ProjectSummary"
     },
     "title": {
      "title": "Title",
      "type": "string"
     },
     "updated_at": {
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "mode",
     "title",
     "goal",
     "start_date",
     "deadline",
     "status",
     "owner_email",
     "members",
     "deleted",
     "created_at",
     "updated_at",
     "summary"
    ],
    "title": "Project",
    "type": "object"
   },
   "ProjectCreate": {
    "properties": {
     "assumptions": {
      "items": {
       "$ref": "#/components/schemas/AssumptionSeed"
      },
      "title": "Assumptions",
      "type": "array"
     },
     "criteria": {
      "items": {
       "$ref": "#/components/schemas/CriterionSeed"
      },
      "title": "Criteria",
      "type": "array"
     },
     "deadline": {
      "format": "date",
      "title": "Deadline",
      "type": "string"
     },
     "goal": {
      "maxLength": 2000,
      "minLength": 1,
      "title": "Goal",
      "type": "string"
     },
     "members": {
      "items": {
       "type": "string"
      },
      "title": "Members",
      "type": "array"
     },
     "mode": {
      "$ref": "#/components/schemas/Mode"
     },
     "start_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "title": {
      "maxLength": 100,
      "minLength": 1,
      "title": "Title",
      "type": "string"
     }
    },
    "required": [
     "mode",
     "title",
     "goal",
     "deadline"
    ],
    "title": "ProjectCreate",
    "type": "object"
   },
...
   "ProjectStatus": {
    "enum": [
     "active",
     "stopped",
     "completed"
    ],
    "title": "ProjectStatus",
    "type": "string"
   },
   "ProjectSummary": {
    "description": "Snapshot of dashboard numbers stored on the project at evaluation time (for the list view).",
    "properties": {
     "deadline_risk": {
      "default": 0,
      "title": "Deadline Risk",
      "type": "integer"
     },
     "drift": {
      "default": 0,
      "title": "Drift",
      "type": "integer"
     },
     "evaluated_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evaluated At"
     },
     "health_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Health Score"
     },
     "unnecessary": {
      "default": 0,
      "title": "Unnecessary",
      "type": "integer"
     },
     "untested": {
      "default": 0,
      "title": "Untested",
      "type": "integer"
     }
    },
    "required": [
     "health_score",
     "drift",
     "unnecessary",
     "deadline_risk",
     "untested",
     "evaluated_at"
    ],
    "title": "ProjectSummary",
    "type": "object"
   },
   "ProjectUpdate": {
    "properties": {
     "deadline": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Deadline"
     },
     "goal": {
      "anyOf": [
       {
        "maxLength": 2000,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Goal"
     },
     "members": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Members"
     },
     "mode": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Mode"
       },
       {
        "type": "null"
       }
      ]
     },
     "start_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Start Date"
     },
     "status": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ProjectStatus"
       },
       {
        "type": "null"
       }
      ]
     },
     "title": {
      "anyOf": [
       {
        "maxLength": 100,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "title": "ProjectUpdate",
    "type": "object"
   },
...
   "Report": {
    "properties": {
     "content": {
      "$ref": "#/components/schemas/ReportContent"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "created_by": {
      "title": "Created By",
      "type": "string"
     },
     "health_score": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Health Score"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "projectId": {
      "title": "Projectid",
      "type": "string"
     }
    },
    "required": [
     "id",
     "projectId",
     "created_at",
     "created_by",
     "model",
     "health_score",
     "content"
    ],
    "title": "Report",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
   "Weights": {
    "properties": {
     "alignment": {
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Alignment",
      "type": "number"
     },
     "schedule": {
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Schedule",
      "type": "number"
     },
     "validation": {
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Validation",
      "type": "number"
     },
     "waste": {
      "maximum": 1.0,
      "minimum": 0.0,
      "title": "Waste",
      "type": "number"
     }
    },
    "required": [
     "alignment",
     "validation",
     "schedule",
     "waste"
    ],
    "title": "Weights",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "PoC Compass API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/modes/{mode_id}": {
   "get": {
    "operationId": "get_mode_api_modes__mode_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "mode_id",
      "required": true,
      "schema": {
       "title": "Mode Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModeDefinition"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Mode",
    "tags": [
     "modes"
    ]
   },
   "put": {
    "operationId": "update_mode_api_modes__mode_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "mode_id",
      "required": true,
      "schema": {
       "title": "Mode Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ModeDefinitionBody"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModeDefinition"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Mode",
    "tags": [
     "modes"
    ]
   }
  },
  "/api/modes/{mode_id}/reset": {
   "post": {
    "operationId": "reset_mode_api_modes__mode_id__reset_post",
    "parameters": [
     {
      "in": "path",
      "name": "mode_id",
      "required": true,
      "schema": {
       "title": "Mode Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ModeDefinition"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Reset Mode",
    "tags": [
     "modes"
    ]
   }
  },
  "/api/projects": {
   "get": {
    "operationId": "list_projects_api_projects_get",
    "parameters": [
     {
      "in": "query",
      "name": "mode",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Mode"
        },
        {
         "type": "null"
        }
       ],
       "title": "Mode"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Project"
         },
         "title": "Response List Projects Api Projects Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Projects",
    "tags": [
     "projects"
    ]
   },
   "post": {
    "operationId": "create_project_api_projects_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Project"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Project",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/projects/{project_id}": {
   "delete": {
    "operationId": "delete_project_api_projects__project_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Project",
    "tags": [
     "projects"
    ]
   },
   "get": {
    "operationId": "get_project_api_projects__project_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Project"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Project",
    "tags": [
     "projects"
    ]
   },
   "patch": {
    "operationId": "update_project_api_projects__project_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Project"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Project",
    "tags": [
     "projects"
    ]
   }
  },
...
  "/api/projects/{project_id}/dashboard": {
   "get": {
    "operationId": "dashboard_api_projects__project_id__dashboard_get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Dashboard"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Dashboard",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/projects/{project_id}/evaluations": {
   "post": {
    "operationId": "evaluate_api_projects__project_id__evaluations_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Evaluation"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Evaluate",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/projects/{project_id}/evaluations/latest": {
   "get": {
    "operationId": "latest_evaluation_api_projects__project_id__evaluations_latest_get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "anyOf": [
          {
           "$ref": "#/components/schemas/Evaluation"
          },
          {
           "type": "null"
          }
         ],
         "title": "Response Latest Evaluation Api Projects  Project Id  Evaluations Latest Get"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Latest Evaluation",
    "tags": [
     "projects"
    ]
   }
  },
...
  "/api/projects/{project_id}/reports": {
   "get": {
    "operationId": "list_reports_api_projects__project_id__reports_get",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Report"
         },
         "title": "Response List Reports Api Projects  Project Id  Reports Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Reports",
    "tags": [
     "projects"
    ]
   },
   "post": {
    "operationId": "create_report_api_projects__project_id__reports_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Report"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Report",
    "tags": [
     "projects"
    ]
   }
  },
  "/api/projects/{project_id}/tasks/extract": {
   "post": {
    "operationId": "extract_tasks_api_projects__project_id__tasks_extract_post",
    "parameters": [
     {
      "in": "path",
      "name": "project_id",
      "required": true,
      "schema": {
       "title": "Project Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExtractRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ExtractResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Extract Tasks",
    "tags": [
     "projects"
    ]
   }
  }
 }
}